refactor: Move palette definition to autoload functions
Allows using palette colors when overwriting the highlight groups. Without autoload functions, it is not possible to retrieve the palette since (i) the `s:palette` variable is not globally accessible, and (ii) the palette may simply not be defined yet by the time the custom highlight groups are given (even if specified as `g:palette`).